requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
distro==1.9.0
Django==5.0.2
django-environ==0.11.2
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
nltk==3.8.1
openai==1.30.1
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.7.1
pydantic_core==2.18.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rsa==4.9
simplejson==3.19.2
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
